backend/requirements.txt
pandas
openpyxl
django-cors-headers